Made accordion accessible
This was done to ensure the accordion menu works well for screen
readers.

There are downsides to this approach. It requires that the control id
matches with the content, creating more for the user of the lib to do.
This makes it a little more fragile. But this is what I know as the best
way to do this for accessibility.

Research from here and other projects.
http://heydonworks.com/practical_aria_examples/#progressive-collapsibles