Changes to borders.
The borders were changed to match with mockups including a border
radius on the whole element. THe overflow:hidden rule on the whole
element was done so that the border radius works correctly on it, other-
wise, the corners of the li child elements will poke out. This could
also be fixed by adding border radiuses to the li elements but I picked
this method because it's less code.