Style the password requirments box
This styles the password requirements box so it's an info box with
just one header line, a list of requirements that can have a check
as active which can be two lines. So the list was made so when
there are two lines, the second line will be indented to line up with
the first.