Validator for password reset
This includes code that can perform validations on regexes defined
through data attributes. It requires the user to write regexes on the
input element that can match up with the line items in a checkbox style
list.

This was done this way to ensure all configuration could be done easily
from within HTML, requiring less to be done in JS.