Grids that breakdown on the middle breakpoint
The difference with this system is that you have to wrap each row in a
`usa-grid`. This is because the nth-child rules work based on the parent
wrapper. So if each row isn't wrapped, then the nth-child rule will be
based on the full wrapper. This means, for instance, that on the grid
page, nth-child(2n) would end up being the second element in the whole
thing(the first 1/2 element) rather then the 3rd (the 2nd 1/2 element).

On the bourbon grid example page, the mobile/medium breakpoint shrinks
everything down to 100%, which is why this works.