V2 grids, remove layouts
We're removing the "Layouts" portion of the page because it's not needed
for MVP. Since Layouts and Grids were both under a top nav of "Layout
System", this makes Grids the top level item and removes "Layout System"
and "Layouts".

Work isn't finished with removing layout-system from structure but is
also not required to have this working.