WIP test on new documentation scheme
This is to show what the new documentation scheme looks like.

So on building this code, I realized we made a mistake by putting such
large top/down paddings on our accordion content. You should be able
to place any content in an accordion.

What happens: if you placed content with hs and ps in an accordion, the
top and bottom padding was overwhelming, because it was the margins of
the h and p tags PLUS the padding in teh accordion. To fix this, I
removed the padding from the accordion and ensured the content we're
putting in there is wrapped in a semantic tag, like <p>.

This brought on another problem: p tags are confined to a max width,
so now the accordion content doesn't go to the full width of the
accordion. You should be able to use p tags in an accordion. This is
also an issue for the lead-text class which uses a p tag.