Fix padding and max width
This includes code to fix the padding so the accordion is in charge
of defining the padding of the element. This ensures it will always
be padded the same no matter the content.

In the future, we should consider not having margin tops on the
headings. This would mean we wouldn't have to clear it in the
accordion.

Also removed the usa-content class as we'll be moving it to be more
localized. We'll add that class back in specific places.