Attempt to fix focus state on accordion hover.
An important rule has to be used because theres another one on
"unstyled" button. We should address this at some point, possibly
now.

The other larger problem is the <ul> on the accordion has a
overflow: hidden on it. This means the the box shadow won't show
up on accordions using the <ul>. Is there a strong reason why we
need overflow:hidden on the accordion <ul>?