_footer.scss uses $medium instead of $medium-screen
The current version of the file generates a final CSS snippet of 

```css
@media screen and (min-width: min-width 600px 6) {
  .width-one-sixth {
    float: left;
    display: block;
    margin-right: 2.35765%;
    width: 14.70196%; }
    .width-one-sixth:last-child {
      margin-right: 0; } }
```

The problematic part is the `min-width: min-width 600px 6`. Changing `$medium` to `$medium-screen` makes it `min-width: 600px`. I admit that I am new to SCSS programming, but that seems like the right thing by analogy to the other calls.