Update package.json hash;
This patch removes the ambiguity around the style, css, scss properties
in the `package.json` hash. There should only be one declaration for styles
in the `package.json` since a user using this file to consume the styles
_should_ have an asset pipeline which consumes the `scss` file and there
would be no need to use the compiled version created on `npm-prepublish`.