Split up JavaScript into component files;
This patch is the start of the work that needs to be done for dependency
management and documentation around it. I am going to leave the
refactoring of this to @msecret as he has some production experience of
doing this with the standards. This work should be a good starting
point for this ticket =>
https://trello.com/c/3J13uLIH/34-as-a-master-builder-i-want-to-be-able-to-know-what-dependencies-are-required-to-successfully-work-with-the-wds-and-i-want-the-op