Add ignorance for dist directory;
This patch adds the dist/ directory ignorance to the project. The dist
directory is used locally to build the package into various consumable
projects. It should never be committed into the project entirely since
it may contain multiple copies of the same files in different
directories. Essentially, this is the landing place for all built files.