Solution for Sass module paths
Use the `bourbon` and `bourbon-neat` npm packages which are the
official packages and will *probably* be more up to date.

Before we were maniuplating the includePaths in the gulp build task
to include the bourbon and neat include paths. This was problematic
because if you installed uswds from another project, you'd also have
to know to set those include paths OR install a secondary tool like
eyeglass.

This solution doesn't use different paths for bourbon and neat but
includes them directly from node_modules. Benefits and negatives of this
approach:
+ npm consumers of uswds can just import it and it will work
- if bourbon or neat change the paths of their module in a future update
things will break.

I don't think we have the perfect solution to this but wanted to use
this solution in the time being as we look at Eyeglass more to see if
it's becoming an industry standard.

More info can be found in #1013. Closes #1013.