Update project ignorance;
This patch updates the git-ignorance around the `dist/` directory. For
the `0.9.0` release we will be including this _compiled and bundled_
directory to help users who clone or download the Git repository and to
avoid confusion around any documentation that references a `dist`
directory that wasn't previously there in the committed files.