Add `dist/` directory;
This patch adds the `dist/` directory compiled as of this commit. This
directory needs to be compiled whenever there are file changes to the
project. As of this patch, the `dist/` directory does not include the
source `scss` or `js` files. It only includes the bundled and compiled
versions of those assets.

This `dist/` directory was compiled with the following command:
`gulp no-gem no-test production build`