Update javascript task;
This patch adds a `defaultStream` and a `minifiedStream` for the
`javascript` task to output both a uswds.js and a uswds.min.js file in
the `dist/` directory. Sourcemaps are also supported via headers and a
separate `uswds.min.js.map` file is created. This file is not pulled
in via a sourcemap comment and must be served via the `X-SourceMap`
header.

Source: http://www.html5rocks.com/en/tutorials/developertools/sourcemaps/#toc-howwork
```
If you don't like the idea of the weird comment you can alternatively
set a special header on your compiled JavaScript file:

X-SourceMap: /path/to/file.js.map
```