Update sass task;
This patch adds a `defaultStream` and a `minifiedStream` for the
`sass` task to output both a uswds.css and a uswds.min.css file in
the `dist/` directory. Sourcemaps are also supported via headers and a
separate `uswds.min.css.map` file is created. This file is not pulled
in via a sourcemap comment and must be served via the `X-SourceMap`
header.

Source: http://www.html5rocks.com/en/tutorials/developertools/sourcemaps/#toc-howwork
```
If you don't like the idea of the weird comment you can alternatively
set a special header on your compiled CSS file:

X-SourceMap: /path/to/file.css.map
```