Remove the jekyll gem dependency;
This patch removes the Jekyll gem dependency from the `docs/` site.
There are a few moving parts to this process. Because we are no longer
using the gem all files need to be copied over to their respective
directories. A list of files and directories that are generated using
`gulp website:serve` are in the `.gitignore` file. These files will be
ignored from version control since they already exist in many different
variations.