Have to remove precommit because system is broken
We have `.scss-lint` in npmignore, so if you attempt to install
this branch, or the uswds repo from github, you won't get this file.
You also won't get the `src/stylesheets/libs/` directory which means
you'll have to go into the uswds folder and run `npm install` and
`gulp copy-vendor-sass` to get uswds working. Because of this hook
`npm install` will automatically run `npm run build` lacking the
`no-test` flag which means you'll try and run the scss-lint gulp
task with no `.scss-lint` file.

We should really put this branch into a release, or remove this prepublish
hook which is so problematic, or add the sass libs to npm, so I'm not
completely blocked on using the standards for cloud.gov.