Move all image/fonts to use vars for path
To make cg-style easier to work with in webpack (which parses all
`url` statments in CSS, actually finds the images/fonts for them,
and copies them) the path to image and font assets should be a
configurable var. This is standard practice seen in both bootstrap
and Salesforce Lightning Design.

This builds on #952-override_variables as none of this would work
without variable overriding.

Fixes #1007.