Dismiss button and dismiss alert
For my current project we wanted a way to dismiss alerts. The solution
we found was to include a form in the alert with a button. So, we
automatically got a dismiss button. However, we may want to include an
inverse as well for accessibility and different background color
applications.