no list accordions
Fixes #1136

I am not sure what the conversation around placing each accordion
within a list item was; however, it feels less than semantic to have a
list item containing a button and a div…imagine a text document with a
list item containing an entire chapter of a book, for example.

I’m also curious if a more suitable tag (say, section) would be more
proper, given the specification of a div and section:

http://www.w3.org/TR/html5/grouping-content.html#the-div-element - The
div element has no special meaning at all. … Authors are strongly
encouraged to view the div element as an element of last resort, for
when no other element is suitable.

http://www.w3.org/TR/html5/sections.html#the-section-element - The
section element represents a generic section of a document or
application.

Changing to a section tag could get us toward a no JavaScript solution
to the accordions (#1122 and #1121) - basically, add the buttons and
collapse the content if JS is present.

Could result in “flashing” content though prior to JS kicking in.