Removed copied src sass
The src sass is not being copied over into the `docs` folder anymore.

This was done by using the `load_paths` option in the jekyll config
which allows to set a path outside of the docs dir. With this, one load
path is set to src/stylesheets. The references in homepage.scss and
styleguide.scss are both changes to "all" so they import directly from
src.

This means the website always consumes the version of the standards in
the src directory.

TODO: This doesn't use the WDS in a way that a user of them would as
it's a bit of a hack around things.