Reverting the changes to the styleguide for now;
This patch reverts @tysongach's work on removing bourbon prefixes. The
reason for this revert is that these styles are a part of the
documentation site and not the actual package. Once we update to version
5.0 of Bourbon, we will need to revisit this. But because the
autoprefixer runs in the `gulp` asset pipeline and not in the `jekyll *`
pipeline, implementing this now is non-trivial.