Change paths and use normalize.scss
So node-sass can import .css files but ruby sass cannot, so I want
to ditch the annoying normalize lib for the scss version.

I'm using the straight node_module paths, which isn't the perfect
solution. Something better but would be much more complicated to
get working on sites that use the standard is to set the includePath
in the sass build path. I'm hesitant to do this now because node-sass
and ruby-sass use these differently so I'm not sure how we'll use them
both.

TODO
- Remove unnecessary build steps.