[Website, Label, Alerts, Links] Background color and text color mixin and classes
## Description

* Fixes #1158
* Fixes #1242
* Groundwork for #1239

* Adds definition for links on a “light” background color.
* Adds `background-color` to `/src/core/_utilities.scss` using Sass
conditionals instead of individual `@mixins` like #1129.
* Adds import of **new** `/src/stylesheets/core/_colors.scss` file.
* Deprecates most `.usa-color-*` classes in
`/docs/doc_assets/styleguide.scss` as they are promoted to a
`_colors.scss` (save `.usa-color-white`) - classes not removed from
`styleduide.scss` with this iteration of the PR as I am not sure where
they are used.

## Additional information

* This iteration is limited to effects on the “Visual style” page only.
Will await feedback on changes and desire before spreading out to other
areas of the Standards.
* This uses `color-gray-dar` for dark links on light backgrounds -
contrast of ~7.98:1. This choice was based on an assumption that light
backgrounds would mirror more the dark background solution (matching
the plain text color more) as opposed to the default link colors.
* The single `@mixin` allows for more detailed definition and
combinations (if we choose a different link text color, we can add a
conditional, which would automatically update results for those using
the classes or `@mixin`).
* It does not, at this point, address the desire to allow developers to
use the `@mixin` and set specific background and font colors; however,
in those cases, I believe the recommendation would be not to use the
`@mixin` or classes…just straight `background-color` CSS.

Tagging the parties from the previous PR: @maya, @msecret, @carodew