[Footer]: Fix social icon hit area for mobile
## Description

Fixes #346

## Additional information

* This PR uses `display: inline-block` to allow setting height and
width to an inline element.
* Uses 44 pixels for both height and width per the Apple iOS HIG
* Concerns with `text-align: right` - could result in thumb hitting to
the right of the hit area - actually hitting the icon to the right of
the actual target. `text-align: center` used instead.
* Using `text-align: center` causes the right alignment of the icons to
be off when compared “Agency Contact Center” text. `position: relative;
left: 13px` used in order to compensate. Slight concern that this could
cause a horizontal scroll bar to appear on mobile - though, given the
overall alignment of the `social-links` container - concern may be moot.