Re-engineered social links
## Description

Alternative #1248

## Additional information

#1248 updates the size of the social link icons, and adds a hit area
around the links to make them more mobile friendly. Visually, the links
are smaller; however, the active area of the link is larger (~45px
square).

Things I noticed from creating #1248:

- Changing icon sizes (#706) requires manually changing the height and
width per link - in two areas (once for the `svg` element and the other
for the `image` element).
- Adding a new icon (LinkedIn, for example), requires adding 6 new
lines to define the link.
- The CSS definition assumes all social links need to be held within a
container `.usa-social-links`, which has no definition outside of
defining each `a` element within the container.

This PR:

1. Re-engineers social icons as background images.
2. Removes dependency on `.usa-social-links` container.
3. Removes containing `div`.
4. Adds `.usa-social-link` class for defining global attributes (hit
area, for example); allowing any link to become a social link.
5. Adds `.facebook`, `.twitter`, `.youtube`, and `.rss` classes, which
must be used in conjunction with `.usa-social-link` to take effect.
6. Allows for introduction of more background for social media links by
extenders, without having to define dimensions and layouts.

This is a branch from a branch (could not get the image updates from
#1248 in here - possibly a knowledge gap on my part); therefore, if we
would like to go this route, I should be able to merge this into the
other to update #1248…at which point, this PR can be closed without
merging.