Ensure opened link fills to bottom of header
When a menu is opened on the header, the blue backgrounded link won't
expand all the way to the bottom of the header, leaving an awkward space
between the two. This modifies the spacing of the button and in turn,
the button's graphical arrow so that the opened menu fills it's vertical
space.

As described here:
https://github.com/18F/web-design-standards/issues/1753#issue-211255604