Move styleguide elements out of rendering
SO not the buttons are rendered as just themselves and no wrapping
just for the styleguide. I experimented with this because it might
make sense for the components to have more accurate code samples
like this, and it also allows for the more styleguide aspects of
the viewing the components to be in the actual styleguide code,
on docs site, rather then here.