Use a framed fractal view to showcase component
To show a representation of the element that's closer to what the
design calls for, I've used the `framed` fractal preview, which
wraps the component in a grid, ensuring it appears at a smaller
width.

This was done rather then adding a max-width to the component,
which would break a lot of implementations of it which require
the alert to span the whole width of it's container.

Another possibility which could be explored would be to add a max
width to the body content.