Add correct bar variable color
I did this using a second map. I tried to find a way to do it by
being dynamic with variables, but Sass doesn't like dynamic vars,
so I think two maps might be the only way. That being said, happy
to look at other ideas on how to do it.