Lower specificity of form
By removing qualifying types that I tested and weren't required,
at least for the forms on the form template components.

Also ensured the form sectino is wrapped in an actual form, because
that's generally how it will be used on sites

Also ensured the form sectino is wrapped in an actual form, because
that's generally how it will be used on sites.