Remove text color and decoration
Remove text color and decoration from usa-footer-return-to-top section.

Let the styles for the link element do their thing :)

### Reasoning
I'm wanting to put some other content into the `usa-footer-return-to-top ` div that is not a link and therefore don't want the text to be blue and underlined. I'm thinking of using this return to top section as a page footer with meta information like last updated date and a link to take you to GitHub for editing.