Close accordion on sidenav close
When tabbing through the mobile side nav and opening an accordion, it seems like the accordion should either close when you tab to the next top-level item, or the side nav is closed.
The current implementation closes the accordion when you tab to an item not in the <ul> of links, which is a weird in-between option. This commit keeps the accordion open until the side nav is fully closed.